Updating serialization for RegisterFirstMessage
